requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
gunicorn==21.2.0
qrcode==7.4.2
opencv-python-headless==4.8.1.78
tensorflow==2.13.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
Pillow==10.0.0
fpdf==1.7.2